changed some of the plot structure
also added normalization into the task1_notebook, but the results where the same except for the neural network model which actually fit better into the training data but it led to lower predictions for the testing data. I decided to just comment it out,and if someone wants to add normalization using minmax then just uncomment that block.

In the Neco notebook submission, I changed some of the colors to make it have consistensy with the features, also I made the histograms over week and days into boxplots, as I think they tell more.